requirements.txt
certifi==2020.6.20
imutils==0.5.3
numpy==1.19.4
olefile==0.46
opencv-contrib-python==4.4.0.46
opencv-python==4.4.0.46
wincertstore==0.2
Pillow==7.2.0
pygubu==0.10.2
